[Tour Manager]  - Moves TourProgressManager to tour_manager module
