+++ app/build.gradle
@@ -49,6 +49,7 @@ dependencies {
     implementation project(':splash')
     implementation project(':welcome')
     implementation project(':tours')
+    implementation project(':tour_manager')
     implementation project(':info')
     implementation project(':audio')
     implementation project(':map')
