+++ map/build.gradle
@@ -31,4 +31,5 @@ dependencies {
     implementation libs.arch_navigation_ui
     implementation libs.glide
     implementation libs.view_indicator
+    implementation project(':tour_manager')
 }
