+++ tour_manager/build.gradle
@@ -1,0 +1,10 @@
+apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-kapt'
+apply plugin: 'androidx.navigation.safeargs'
+
+dependencies {
+    kapt libs.dagger_compiler
+    kapt libs.dagger_android_compiler
+    implementation project(':db')
+    implementation libs.rx_kotlin
+}
