+++ tour_manager/src/main/java/edu/artic/tours/manager/TourManagerModule.kt
@@ -1,0 +1,20 @@
+package edu.artic.tours.manager
+
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+/**
+ * @author Sameer Dhakal (Fuzz)
+ */
+@Module
+abstract class TourManagerModule {
+    @Module
+    companion object {
+
+        @JvmStatic
+        @Provides
+        @Singleton
+        fun tourProgressManager(): TourProgressManager = TourProgressManager()
+    }
+}
